Draft app harness package
This is the generic bit of an application harness -- (tekton)
pipelines for delivering an applicaion into the cluster.

There's a companion package, which is the pieces to specialise the
harness for a specific app. The idea is you stamp out one of this
latter package for each app you bring on board the cluster.