Write a script to instantiate the app template
It's painful to do all the kpt cfg setters and substs for this, and I
wanted to demonstrate how (and why) you can use programs to generate
YAMLs with kpt.

This is just the script (and the template YAMLs); it needs a
Dockerfile and some infrastructure to make the image that'll be used
via kpt.