Change how ingress rewrite works
The backend stanza controls how a request is dispatched to a
service. In general the backend service itself may not know how this
routing is done, and serve hard-coded paths. E.g., if your backend
path is /cuttlefacts/, then your backend service must know that all
the paths requested will start with /cuttlefacts/.

The rewrite annotation (for the nginx-ingress-controller) rewrites
paths so the backend service doesn't need to know the routing. But I
was using it in an outdated way -- it now should refer to the capture
group from the path.

I changed the field in app.yaml to be prefix, to be clearer about the
role it takes.